feat(request-store): Implement script to retrieve webpage content and store in file
Developed a Node.js script that utilizes the 'request' module to send a GET request to the specified URL and pipes the response to a writable stream created with the 'fs' module. This enables the script to efficiently retrieve the contents of a webpage and store them in a file specified by the user.