feat: Implement script to print characters of a Star Wars movie
Implemented a Node.js script that retrieves and prints the characters of a Star Wars movie specified by the movie ID provided as a command line argument. Utilized the 'request' module to make HTTP requests to the Star Wars API endpoint. Parsed the response body as JSON to access the movie data. Mapped over the characters array of the movie to retrieve character details, creating promises for each character to handle the request. Used Promise.all() to handle all promises concurrently and wait for all requests to complete before printing the character names joined by a newline character.